Added Cobertura tool for code coverage for 'fiiadmis-service'.
I added a new run maven configuration with goal 'cobertura:cobertura'
and I set the base directory (eclipse). By default, it runs the unit
tests before the report.

The report will be generated in the 'build' directory:
.../git/fii-admis/FII-Admis-RESTService/target/site/cobertura/index.html

Of course, for now the code coverage percent is 0%, but the report looks
pretty great :)